v0.8.8: IPv6 geo support, DB1.IPV6 embedded, DB5.IPV6 upgrade, matrix/overlay/crypto modules
- Replace IP2Location IPv4-only DB1 with DB1.IPV6 (2MB, IPv4+IPv6 country-level)
- Remove embedded DB11; city-level now via geo-upgrade (DB5.IPV6, ~34MB)
- geo-upgrade: token-based IP2Location download + release asset fallback
- Fix 'IPv6 address missing in IPv4 BIN' errors in topo rendering
- GeoInfo Latitude/Longitude upgraded to float64
- cleanField filters IPv6-in-IPv4 error messages
- Show 'run: clawnet geo-upgrade' hint when location unknown
- Add matrix discovery, overlay transport, crypto (edwards/NaCl) modules
- Dev mode build tag separation (dev.go / dev_release.go)
- Update TODO.md: check completed items, remove WireGuard section
- .gitignore: protect IP2GEOTOKEN, .ip2geotoken secrets

+++ clawnet-cli/internal/geo/centroids.go
@@ -3,7 +3,7 @@ package geo
 // countryCentroids maps ISO 3166-1 alpha-2 country codes to approximate
 // geographic centroids (latitude, longitude). Used as a fallback when
 // IP2Location DB1 returns 0,0 coordinates (country-only database).
-var countryCentroids = map[string][2]float32{
+var countryCentroids = map[string][2]float64{
 	"AD": {42.5, 1.5},
 	"AE": {24.0, 54.0},
 	"AF": {33.0, 65.0},
@@ -223,7 +223,7 @@ var countryCentroids = map[string][2]float32{
 
 // CountryCentroid returns the approximate centroid for an ISO 3166-1 alpha-2
 // country code. Returns (0, 0, false) if the country is not found.
-func CountryCentroid(cc string) (lat, lon float32, ok bool) {
+func CountryCentroid(cc string) (lat, lon float64, ok bool) {
 	c, found := countryCentroids[cc]
 	if !found {
 		return 0, 0, false
